requirements.txt
Flask==3.0.0
yfinance==0.2.33
pandas==2.1.4
numpy==1.26.2
flask-cors==4.0.0
tensorflow>=2.15.0
scikit-learn>=1.3.0
newsapi-python>=0.2.7
textblob>=0.17.1
